+++ ui/packages/manage/regional-content/RegionalContentList.tsx
@@ -6,13 +6,16 @@ import { RegionalContentDto } from "@fireback/manage/sdk/abac/RegionalContentDto
 import { RegionalContentNavigation } from "@fireback/ui-core/sdk/navigation/AbacNavigation";
 import { columns } from "./RegionalContentColumns";
 import { strings } from "./strings/translations";
+import { createUdfBrowseQueryHook } from "@fireback/ui-core/hooks/useUdfBrowseQuery";
 export const RegionalContentList = () => {
   const s = useS(strings);
   return (
     <>
       <CommonListManager
         columns={columns(s)}
-        queryHook={useRegionalContentBrowseActionQuery}
+        queryHook={createUdfBrowseQueryHook(
+          useRegionalContentBrowseActionQuery,
+        )}
         uniqueIdHrefHandler={(uniqueId: string) =>
           RegionalContentNavigation.single(uniqueId)
         }
