+++ ui/packages/selfservice/public-join-keys/PublicJoinKeyList.tsx
@@ -1,8 +1,8 @@
-import { CommonListManager } from "@fireback/ui-core/components/entity-manager/CommonListManager";
-import { usePublicJoinKeyBrowseActionQuery } from "@fireback/selfservice/sdk/abac/PublicJoinKeyBrowseAction";
 import { usePublicJoinKeyAwareDeleteAction } from "@fireback/selfservice/sdk/abac/PublicJoinKeyAwareDeleteAction";
+import { usePublicJoinKeyBrowseActionQuery } from "@fireback/selfservice/sdk/abac/PublicJoinKeyBrowseAction";
+import { CommonListManager } from "@fireback/ui-core/components/entity-manager/CommonListManager";
 import { useS } from "@fireback/ui-core/hooks/useS";
-import { PublicJoinKeyDto } from "@fireback/selfservice/sdk/abac/PublicJoinKeyDto";
+import { createUdfBrowseQueryHook } from "@fireback/ui-core/hooks/useUdfBrowseQuery";
 import { PublicJoinKeyNavigation } from "@fireback/ui-core/sdk/navigation/AbacNavigation";
 import { columns } from "./PublicJoinKeyColumns";
 import { strings } from "./strings/translations";
@@ -14,7 +14,7 @@ export const PublicJoinKeyList = () => {
     <>
       <CommonListManager
         columns={columns(s)}
-        queryHook={usePublicJoinKeyBrowseActionQuery}
+        queryHook={createUdfBrowseQueryHook(usePublicJoinKeyBrowseActionQuery)}
         uniqueIdHrefHandler={(uniqueId: string) =>
           PublicJoinKeyNavigation.single(uniqueId)
         }
